+++ smallworld/utils.py
@@ -217,6 +217,11 @@ def analyze(state: T) -> None:
     """
 
     for name in analyses.__all__:
-        module = getattr(analyses, name)()
-        if isinstance(module, analyses.Analysis):
-            module.run(state)
+        module: typing.Type = getattr(analyses, name)
+        if issubclass(module, analyses.Filter) and module is not analyses.Filter:
+            module().activate()
+
+    for name in analyses.__all__:
+        module = getattr(analyses, name)
+        if issubclass(module, analyses.Analysis) and module is not analyses.Analysis:
+            module().run(state)