+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - `AngrEmulator` based on angr symbolic execution.
 - `AngrNWBTAnalysis` unused value analysis using angr.
+- `Filter` - analyses that simply listen to the hint stream.
 
 ### Changed
 - `State.map()` automatically selects names for mapped objects when not